backend/requirements.txt
fastapi
uvicorn
motor
beanie
passlib[argon2]
python-multipart
python-jose
python-dotenv

